Sécurisation des routes non get

+++ backend/models/comment.models.js
@@ -1,18 +1,14 @@
-class Comment {
+module.exports = class Comment {
   static create(callback, params) {
-    let sql = `INSERT INTO comments(message,author,post) VALUES ( ?
-    , ?, ?)`;
+    let sql = `INSERT INTO comments(message,author,post) VALUES (?, ?, ?)`;
     callback(sql, params);
   }
   static update(callback, params) {
-    let sql = `UPDATE comments SET message = ?
-        WHERE id= ?
-        AND author=?`;
+    let sql = `UPDATE comments SET message = ? WHERE (id= ? AND author= ?)`;
     callback(sql, params);
   }
   static delete(callback, params) {
-    let sql = `DELETE FROM comments WHERE id=?
-       AND author=?`;
+    let sql = `DELETE FROM comments WHERE (id= ? AND author= ?) OR (SELECT permissions FROM users WHERE id= ?) = 1`;
     callback(sql, params);
   }
-}
+};
